refactor: update bounty hooks and settings components
- Replaced the use of `useBounty` with `useBountyCRUD` in bounty settings and creation pages for improved functionality.
- Cleaned up the `useBounty` hook by removing unused code and simplifying its structure.
- Enhanced the settings page by integrating the new hook for better state management.
